building-with-llms-main/requirements.txt
# Core dependencies
anthropic>=0.18.1
openai>=1.12.0
google-cloud-aiplatform>=1.42.1
azure-ai-ml>=1.12.1

# Web utilities
aiohttp>=3.9.3
beautifulsoup4>=4.12.3
duckduckgo-search>=4.4.3

# Async support
asyncio>=3.4.3

# Type hints
typing-extensions>=4.9.0

# Environment variables
python-dotenv>=1.0.1

# Testing
pytest>=8.0.2
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0

# Development
black>=24.2.0
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0 